Move LvCoordinator to distinct package
