+++ src/main/scala/edu/ie3/osmogrid/lv/coordinator/LvCoordinator.scala
@@ -1,70 +1,29 @@
 /*
- * © 2021. TU Dortmund University,
+ * © 2022. TU Dortmund University,
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.osmogrid.lv
+package edu.ie3.osmogrid.lv.coordinator
 
-import akka.actor.typed.{ActorRef, Behavior, PostStop, SupervisorStrategy}
-import akka.actor.typed.scaladsl.{ActorContext, Behaviors, Routers}
+import akka.actor.typed.scaladsl.{ActorContext, Behaviors}
+import akka.actor.typed.{ActorRef, Behavior, PostStop}
 import edu.ie3.datamodel.models.input.container.SubGridContainer
 import edu.ie3.osmogrid.ActorStopSupport
 import edu.ie3.osmogrid.cfg.OsmoGridConfig
-import edu.ie3.osmogrid.cfg.OsmoGridConfig.Generation.Lv
 import edu.ie3.osmogrid.exception.RequestFailedException
 import edu.ie3.osmogrid.io.input.InputDataProvider
 import edu.ie3.osmogrid.io.input.InputDataProvider.{ReqAssetTypes, ReqOsm}
-import edu.ie3.osmogrid.lv.LvCoordinator.{Request, cleanUp}
-import edu.ie3.osmogrid.lv.LvGridGenerator
+import edu.ie3.osmogrid.lv.LvRegionCoordinator
 import edu.ie3.osmogrid.model.{OsmoGridModel, PbfFilter}
 import org.slf4j.Logger
 
 import java.util.UUID
-import javax.xml.transform.dom.DOMSource
 import scala.util.{Failure, Success, Try}
 
 /** Actor to take care of the overall generation process for low voltage grids
   */
 object LvCoordinator extends ActorStopSupport[Request] {
-  sealed trait Request
-  object ReqLvGrids extends Request
-  final case class StartGeneration(
-      lvConfig: OsmoGridConfig.Generation.Lv,
-      regionCoordinator: ActorRef[LvRegionCoordinator.Request]
-  ) extends Request
-  object Terminate extends Request
-
-  /** Container class for message adapters
-    *
-    * @param inputDataProvider
-    *   Message adapter for responses from [[InputDataProvider]]
-    * @param regionCoordinator
-    *   Message adapter for responses from [[LvRegionCoordinator]]
-    */
-  private final case class MessageAdapters(
-      inputDataProvider: ActorRef[InputDataProvider.Response],
-      regionCoordinator: ActorRef[LvRegionCoordinator.Response]
-  )
-
-  private object MessageAdapters {
-    final case class WrappedInputDataResponse(
-        response: InputDataProvider.Response
-    ) extends Request
-
-    final case class WrappedRegionResponse(
-        response: LvRegionCoordinator.Response
-    ) extends Request
-  }
-
-  sealed trait Response
-
-  /** Replying the generated low voltage grids
-    *
-    * @param grids
-    *   Collection of low voltage grids
-    */
-  final case class RepLvGrids(grids: Seq[SubGridContainer]) extends Response
 
   /** Build a [[LvCoordinator]] with given additional information
     *
@@ -97,24 +56,6 @@ object LvCoordinator extends ActorStopSupport[Request] {
       IdleData(cfg, inputDataProvider, runGuardian, messageAdapters)
     )
   }
-
-  /** State data for orientation of the actor
-    *
-    * @param cfg
-    *   Config for the generation process
-    * @param inputDataProvider
-    *   Reference to the [[InputDataProvider]]
-    * @param runGuardian
-    *   Reference to the [[RunGuardian]] to report to
-    * @param msgAdapters
-    *   Collection of all necessary message adapters
-    */
-  private final case class IdleData(
-      cfg: OsmoGridConfig.Generation.Lv,
-      inputDataProvider: ActorRef[InputDataProvider.Request],
-      runGuardian: ActorRef[Response],
-      msgAdapters: MessageAdapters
-  )
 
   /** Idle state to receive any kind of [[Request]]
     *
@@ -160,68 +101,6 @@ object LvCoordinator extends ActorStopSupport[Request] {
     .receiveSignal { case (ctx, PostStop) =>
       postStopCleanUp(ctx.log)
     }
-
-  /** State data to describe the actor's orientation while awaiting replies
-    *
-    * @param osmData
-    *   Current state of information for open street maps data
-    * @param assetInformation
-    *   Current state of information for asset data
-    * @param cfg
-    *   Config for the generation process
-    * @param msgAdapters
-    *   Collection of available message adapters
-    * @param guardian
-    *   Reference to the guardian actor
-    */
-  private final case class AwaitingData(
-      osmData: Option[OsmoGridModel],
-      assetInformation: Option[InputDataProvider.AssetInformation],
-      cfg: OsmoGridConfig.Generation.Lv,
-      msgAdapters: MessageAdapters,
-      guardian: ActorRef[Response]
-  ) {
-    def registerResponse(
-        response: InputDataProvider.Response,
-        log: Logger
-    ): Try[AwaitingData] = response match {
-      case InputDataProvider.RepOsm(_, osmModel) =>
-        log.debug(s"Received OSM data.")
-        Success(copy(osmData = Some(osmModel)))
-      case InputDataProvider.RepAssetTypes(assetInformation) =>
-        log.debug(s"Received asset information.")
-        Success(
-          copy(assetInformation = Some(assetInformation))
-        )
-      /* Those states correspond to failed operation */
-      case _: InputDataProvider.InvalidOsmRequest =>
-        Failure(
-          RequestFailedException(
-            "The sent OSM data request was invalid. Stop generation."
-          )
-        )
-      case InputDataProvider.OsmReadFailed(reason) =>
-        Failure(
-          RequestFailedException(
-            "The requested OSM data cannot be read. Stop generation."
-          )
-        )
-    }
-
-  }
-
-  private object AwaitingData {
-    def empty(
-        coordinatorData: IdleData,
-        guardian: ActorRef[Response]
-    ): AwaitingData = AwaitingData(
-      None,
-      None,
-      coordinatorData.cfg,
-      coordinatorData.msgAdapters,
-      coordinatorData.runGuardian
-    )
-  }
 
   /** Await incoming input data and register it
     *
